Split scoring.{h,cc}
I've tried to simplify scoring.{h,cc} and split it. What I couldn't
split I let it in bscores.h

Now you still use scoring.h header the same way, it just includes a
bunch of those new files.
