+++ opencog/learning/moses/scoring/CMakeLists.txt
@@ -1,6 +1,11 @@
 #install header files
 INSTALL(FILES
 	scoring.h
+    scoring_base.h
+    discriminating_bscore.h
+    precision_bscore.h
+    instance_scorer.h
+    bscores.h
 
 	DESTINATION
